+++ src/main/java/com/ekart/services/ProductService.java
@@ -4,7 +4,6 @@ package com.ekart.services;
 import com.ekart.entity.Product;
 import com.ekart.response.ProductResponse;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface ProductService {
@@ -13,7 +12,7 @@ public interface ProductService {
 
     List<ProductResponse> getProductsByCategory(String category);
 
-    Product getProductById(BigInteger id);
+    Product getProductById(Long id);
 
     Product saveProduct(Product product);
 }
